gitgudrequirements.txt
pandas
azure-iot-device
azure-storage-blob
